+++ ui/src/dashboards/containers/DashboardPage.tsx
@@ -25,6 +25,7 @@ import idNormalizer, {TYPE_ID} from 'src/normalizers/id'
 import {millisecondTimeRange} from 'src/dashboards/utils/time'
 import {getDeep} from 'src/utils/wrappers'
 import {updateDashboardLinks} from 'src/dashboards/utils/dashboardSwitcherLinks'
+import AutoRefresh from 'src/utils/AutoRefresh'
 
 // APIs
 import {loadDashboardLinks} from 'src/dashboards/apis'
@@ -104,93 +105,76 @@ interface Props extends ManualRefreshProps, WithRouterProps {
   deleteDashboardCellAsync: typeof dashboardActions.deleteDashboardCellAsync
   templateVariableLocalSelected: typeof dashboardActions.templateVariableLocalSelected
   getDashboardWithTemplatesAsync: typeof dashboardActions.getDashboardWithTemplatesAsync
-  rehydrateNestedTemplatesAsync: typeof dashboardActions.rehydrateNestedTemplatesAsync
+  rehydrateTemplatesAsync: typeof dashboardActions.rehydrateTemplatesAsync
   updateTemplateQueryParams: typeof dashboardActions.updateTemplateQueryParams
   updateQueryParams: typeof dashboardActions.updateQueryParams
 }
 
 interface State {
-  isEditMode: boolean
-  selectedCell: DashboardsModels.Cell | null
   scrollTop: number
   windowHeight: number
+  selectedCell: DashboardsModels.Cell | null
   dashboardLinks: DashboardsModels.DashboardSwitcherLinks
 }
 
 @ErrorHandling
 class DashboardPage extends Component<Props, State> {
-  private intervalID: number
-
   public constructor(props: Props) {
     super(props)
 
     this.state = {
-      isEditMode: false,
-      selectedCell: null,
       scrollTop: 0,
+      selectedCell: null,
       windowHeight: window.innerHeight,
       dashboardLinks: EMPTY_LINKS,
     }
   }
 
   public async componentDidMount() {
-    const {source, getAnnotationsAsync, timeRange, autoRefresh} = this.props
+    const {autoRefresh} = this.props
 
-    const annotationRange = millisecondTimeRange(timeRange)
-    getAnnotationsAsync(source.links.annotations, annotationRange)
-
-    if (autoRefresh) {
-      this.intervalID = window.setInterval(() => {
-        getAnnotationsAsync(source.links.annotations, annotationRange)
-      }, autoRefresh)
-    }
+    AutoRefresh.poll(autoRefresh)
+    AutoRefresh.subscribe(this.fetchAnnotations)
 
     window.addEventListener('resize', this.handleWindowResize, true)
 
     await this.getDashboard()
 
+    this.fetchAnnotations()
     this.getDashboardLinks()
   }
 
-  public componentWillReceiveProps(nextProps: Props) {
-    const {source, getAnnotationsAsync, timeRange} = this.props
-    if (this.props.autoRefresh !== nextProps.autoRefresh) {
-      clearInterval(this.intervalID)
-      this.intervalID = null
-      const annotationRange = millisecondTimeRange(timeRange)
-      if (nextProps.autoRefresh) {
-        this.intervalID = window.setInterval(() => {
-          getAnnotationsAsync(source.links.annotations, annotationRange)
-        }, nextProps.autoRefresh)
-      }
-    }
+  public fetchAnnotations = () => {
+    const {source, timeRange, getAnnotationsAsync} = this.props
+    const rangeMs = millisecondTimeRange(timeRange)
+    getAnnotationsAsync(source.links.annotations, rangeMs)
   }
 
   public componentDidUpdate(prevProps: Props) {
+    const {dashboard, autoRefresh} = this.props
+
     const prevPath = getDeep(prevProps.location, 'pathname', null)
     const thisPath = getDeep(this.props.location, 'pathname', null)
 
-    const templates = getDeep<TempVarsModels.Template[]>(
-      this.props.dashboard,
-      'templates',
-      []
-    ).map(t => t.tempVar)
-    const prevTemplates = getDeep<TempVarsModels.Template[]>(
-      prevProps.dashboard,
-      'templates',
-      []
-    ).map(t => t.tempVar)
-    const isTemplateDeleted: boolean =
-      _.intersection(templates, prevTemplates).length !== prevTemplates.length
+    const templates = this.parseTempVar(dashboard)
+    const prevTemplates = this.parseTempVar(prevProps.dashboard)
+
+    const intersection = _.intersection(templates, prevTemplates)
+    const isTemplateDeleted = intersection.length !== prevTemplates.length
 
     if ((prevPath && thisPath && prevPath !== thisPath) || isTemplateDeleted) {
       this.getDashboard()
     }
+
+    if (autoRefresh !== prevProps.autoRefresh) {
+      AutoRefresh.poll(autoRefresh)
+    }
   }
 
   public componentWillUnmount() {
-    clearInterval(this.intervalID)
-    this.intervalID = null
+    AutoRefresh.stopPolling()
+    AutoRefresh.unsubscribe(this.fetchAnnotations)
+
     window.removeEventListener('resize', this.handleWindowResize, true)
     this.props.handleDismissEditingAnnotation()
   }
@@ -217,7 +201,6 @@ class DashboardPage extends Component<Props, State> {
       cellQueryStatus,
       thresholdsListType,
       thresholdsListColors,
-
       inPresentationMode,
       handleChooseAutoRefresh,
       handleShowCellEditorOverlay,
@@ -269,7 +252,7 @@ class DashboardPage extends Component<Props, State> {
       templatesIncludingDashTime = []
     }
 
-    const {isEditMode, dashboardLinks} = this.state
+    const {dashboardLinks} = this.state
 
     return (
       <div className="page dashboard-page">
@@ -295,15 +278,12 @@ class DashboardPage extends Component<Props, State> {
         <DashboardHeader
           dashboard={dashboard}
           timeRange={timeRange}
-          isEditMode={isEditMode}
           autoRefresh={autoRefresh}
           isHidden={inPresentationMode}
           onAddCell={this.handleAddCell}
           onManualRefresh={onManualRefresh}
           zoomedTimeRange={zoomedTimeRange}
-          onSave={this.handleRenameDashboard}
-          onCancel={this.handleCancelEditDashboard}
-          onEditDashboard={this.handleEditDashboard}
+          onRenameDashboard={this.handleRenameDashboard}
           dashboardLinks={dashboardLinks}
           activeDashboard={dashboard ? dashboard.name : ''}
           showTemplateControlBar={showTemplateControlBar}
@@ -344,6 +324,12 @@ class DashboardPage extends Component<Props, State> {
         ) : null}
       </div>
     )
+  }
+
+  public parseTempVar(
+    dashboard: DashboardsModels.Dashboard
+  ): TempVarsModels.Template[] {
+    return getDeep(dashboard, 'templates', []).map(t => t.tempVar)
   }
 
   private handleWindowResize = (): void => {
@@ -430,21 +416,12 @@ class DashboardPage extends Component<Props, State> {
     this.props.cloneDashboardCellAsync(dashboard, cell)
   }
 
-  private handleEditDashboard = (): void => {
-    this.setState({isEditMode: true})
-  }
-
-  private handleCancelEditDashboard = (): void => {
-    this.setState({isEditMode: false})
-  }
-
   private handleRenameDashboard = async (name: string): Promise<void> => {
     const {dashboard} = this.props
-    this.setState({isEditMode: false})
-    const newDashboard = {...dashboard, name}
+    const renamedDashboard = {...dashboard, name}
 
-    this.props.updateDashboard(newDashboard)
-    await this.props.putDashboard(newDashboard)
+    this.props.updateDashboard(renamedDashboard)
+    await this.props.putDashboard(renamedDashboard)
     this.updateActiveDashboard()
   }
 
@@ -461,13 +438,11 @@ class DashboardPage extends Component<Props, State> {
       dashboard,
       source,
       templateVariableLocalSelected,
-      updateTemplateQueryParams,
-      rehydrateNestedTemplatesAsync,
+      rehydrateTemplatesAsync,
     } = this.props
 
     templateVariableLocalSelected(dashboard.id, template.id, value)
-    updateTemplateQueryParams(dashboard.id)
-    rehydrateNestedTemplatesAsync(dashboard.id, source)
+    rehydrateTemplatesAsync(dashboard.id, source)
   }
 
   private handleSaveTemplateVariables = async (
@@ -592,7 +567,7 @@ const mdtp = {
   templateVariableLocalSelected: dashboardActions.templateVariableLocalSelected,
   getDashboardWithTemplatesAsync:
     dashboardActions.getDashboardWithTemplatesAsync,
-  rehydrateNestedTemplatesAsync: dashboardActions.rehydrateNestedTemplatesAsync,
+  rehydrateTemplatesAsync: dashboardActions.rehydrateTemplatesAsync,
   updateTemplateQueryParams: dashboardActions.updateTemplateQueryParams,
   updateQueryParams: dashboardActions.updateQueryParams,
   handleChooseAutoRefresh: appActions.setAutoRefresh,