+++ ui/src/shared/components/Gauge.tsx
@@ -1,39 +1,66 @@
 import React, {Component} from 'react'
-import PropTypes from 'prop-types'
 import _ from 'lodash'
 
-import {GAUGE_SPECS} from 'shared/constants/gaugeSpecs'
+import {GAUGE_SPECS} from 'src/shared/constants/gaugeSpecs'
 
 import {
   COLOR_TYPE_MIN,
   COLOR_TYPE_MAX,
   MIN_THRESHOLDS,
-} from 'shared/constants/thresholds'
+} from 'src/shared/constants/thresholds'
+import {MAX_TOLOCALESTRING_VAL} from 'src/dashboards/constants'
 
-import {colorsStringSchema} from 'shared/schemas'
 import {ErrorHandling} from 'src/shared/decorators/errors'
 
+import {ColorString} from 'src/types/colors'
+import {DecimalPlaces} from 'src/types/dashboards'
+
+interface Props {
+  width: string
+  height: string
+  gaugePosition: number
+  colors?: ColorString[]
+  prefix: string
+  suffix: string
+  decimalPlaces: DecimalPlaces
+}
+
 @ErrorHandling
-class Gauge extends Component {
-  constructor(props) {
+class Gauge extends Component<Props> {
+  private canvasRef: React.RefObject<HTMLCanvasElement>
+
+  constructor(props: Props) {
     super(props)
+    this.canvasRef = React.createRef()
   }
 
-  componentDidMount() {
+  public componentDidMount() {
     this.updateCanvas()
   }
 
-  componentDidUpdate() {
+  public componentDidUpdate() {
     this.updateCanvas()
   }
 
-  resetCanvas = (canvas, context) => {
+  public render() {
+    const {width, height} = this.props
+    return (
+      <canvas
+        className="gauge"
+        width={width}
+        height={height}
+        ref={this.canvasRef}
+      />
+    )
+  }
+
+  private resetCanvas = (canvas, context) => {
     context.setTransform(1, 0, 0, 1, 0, 0)
     context.clearRect(0, 0, canvas.width, canvas.height)
   }
 
-  updateCanvas = () => {
-    const canvas = this.canvasRef
+  private updateCanvas = () => {
+    const canvas = this.canvasRef.current
     canvas.width = canvas.height * (canvas.clientWidth / canvas.clientHeight)
     const ctx = canvas.getContext('2d')
 
@@ -74,20 +101,12 @@ class Gauge extends Component {
       )
     }
     this.drawGaugeLines(ctx, centerX, centerY, radius, gradientThickness)
-    this.drawGaugeLabels(
-      ctx,
-      centerX,
-      centerY,
-      radius,
-      gradientThickness,
-      minValue,
-      maxValue
-    )
+    this.drawGaugeLabels(ctx, radius, gradientThickness, minValue, maxValue)
     this.drawGaugeValue(ctx, radius, labelValueFontSize)
     this.drawNeedle(ctx, radius, minValue, maxValue)
   }
 
-  drawGradientGauge = (ctx, xc, yc, r, gradientThickness) => {
+  private drawGradientGauge = (ctx, xc, yc, r, gradientThickness) => {
     const {colors} = this.props
     const sortedColors = _.sortBy(colors, color => Number(color.value))
 
@@ -111,7 +130,7 @@ class Gauge extends Component {
     ctx.stroke()
   }
 
-  drawSegmentedGauge = (
+  private drawSegmentedGauge = (
     ctx,
     xc,
     yc,
@@ -151,7 +170,7 @@ class Gauge extends Component {
     }
   }
 
-  drawGaugeLines = (ctx, xc, yc, radius, gradientThickness) => {
+  private drawGaugeLines = (ctx, xc, yc, radius, gradientThickness) => {
     const {
       degree,
       lineCount,
@@ -160,15 +179,17 @@ class Gauge extends Component {
       lineStrokeLarge,
       tickSizeSmall,
       tickSizeLarge,
+      smallLineCount,
     } = GAUGE_SPECS
 
     const arcStart = Math.PI * 0.75
     const arcLength = Math.PI * 1.5
     const arcStop = arcStart + arcLength
-    const lineSmallCount = lineCount * 5
+    const totalSmallLineCount = lineCount * smallLineCount
+
     const startDegree = degree * 135
     const arcLargeIncrement = arcLength / lineCount
-    const arcSmallIncrement = arcLength / lineSmallCount
+    const arcSmallIncrement = arcLength / totalSmallLineCount
 
     // Semi-circle
     const arcRadius = radius + gradientThickness * 0.8
@@ -203,7 +224,7 @@ class Gauge extends Component {
     }
 
     // Draw Small ticks
-    for (let lt = 0; lt <= lineSmallCount; lt++) {
+    for (let lt = 0; lt <= totalSmallLineCount; lt++) {
       // Rototion before drawing line
       ctx.rotate(startDegree)
       ctx.rotate(lt * arcSmallIncrement)
@@ -222,10 +243,8 @@ class Gauge extends Component {
     }
   }
 
-  drawGaugeLabels = (
+  private drawGaugeLabels = (
     ctx,
-    xc,
-    yc,
     radius,
     gradientThickness,
     minValue,
@@ -238,10 +257,11 @@ class Gauge extends Component {
 
     const gaugeValues = []
     for (let g = minValue; g < maxValue; g += incrementValue) {
-      const roundedValue = Math.round(g * 100) / 100
-      gaugeValues.push(roundedValue.toString())
+      const valueString = this.labelToString(g)
+      gaugeValues.push(valueString)
     }
-    gaugeValues.push((Math.round(maxValue * 100) / 100).toString())
+
+    gaugeValues.push(this.labelToString(maxValue))
 
     const startDegree = degree * 135
     const arcLength = Math.PI * 1.5
@@ -280,24 +300,65 @@ class Gauge extends Component {
     }
   }
 
-  drawGaugeValue = (ctx, radius, labelValueFontSize) => {
+  private drawGaugeValue = (ctx, radius, labelValueFontSize) => {
     const {gaugePosition, prefix, suffix} = this.props
     const {valueColor} = GAUGE_SPECS
-    const maximumFractionDigits = 20
 
     ctx.font = `${labelValueFontSize}px Roboto`
     ctx.fillStyle = valueColor
     ctx.textBaseline = 'middle'
     ctx.textAlign = 'center'
 
+    const valueString = this.valueToString(gaugePosition)
+
     const textY = radius
-    const textContent = `${prefix}${gaugePosition.toLocaleString(undefined, {
-      maximumFractionDigits,
-    })}${suffix}`
+    const textContent = `${prefix}${valueString}${suffix}`
     ctx.fillText(textContent, 0, textY)
   }
 
-  drawNeedle = (ctx, radius, minValue, maxValue) => {
+  private labelToString(value: number): string {
+    const {decimalPlaces} = this.props
+
+    let valueString
+
+    if (decimalPlaces.isEnforced) {
+      const digits = Math.min(decimalPlaces.digits, MAX_TOLOCALESTRING_VAL)
+      valueString = value.toLocaleString(undefined, {
+        minimumFractionDigits: 0,
+        maximumFractionDigits: digits,
+      })
+    } else {
+      valueString = value.toLocaleString(undefined, {
+        minimumFractionDigits: 0,
+        maximumFractionDigits: MAX_TOLOCALESTRING_VAL,
+      })
+    }
+
+    return valueString
+  }
+
+  private valueToString(value: number): string {
+    const {decimalPlaces} = this.props
+
+    let valueString
+
+    if (decimalPlaces.isEnforced) {
+      const digits = Math.min(decimalPlaces.digits, MAX_TOLOCALESTRING_VAL)
+      valueString = value.toLocaleString(undefined, {
+        minimumFractionDigits: digits,
+        maximumFractionDigits: digits,
+      })
+    } else {
+      valueString = value.toLocaleString(undefined, {
+        minimumFractionDigits: 0,
+        maximumFractionDigits: MAX_TOLOCALESTRING_VAL,
+      })
+    }
+
+    return valueString
+  }
+
+  private drawNeedle = (ctx, radius, minValue, maxValue) => {
     const {gaugePosition} = this.props
     const {degree, needleColor0, needleColor1} = GAUGE_SPECS
     const arcDistance = Math.PI * 1.5
@@ -318,29 +379,6 @@ class Gauge extends Component {
     ctx.lineTo(10, 0)
     ctx.fill()
   }
-
-  render() {
-    const {width, height} = this.props
-    return (
-      <canvas
-        className="gauge"
-        width={width}
-        height={height}
-        ref={r => (this.canvasRef = r)}
-      />
-    )
-  }
-}
-
-const {number, string} = PropTypes
-
-Gauge.propTypes = {
-  width: string.isRequired,
-  height: string.isRequired,
-  gaugePosition: number.isRequired,
-  colors: colorsStringSchema.isRequired,
-  prefix: string.isRequired,
-  suffix: string.isRequired,
 }
 
 export default Gauge