+++ ui/src/shared/components/GaugeChart.tsx
@@ -1,7 +1,7 @@
 import React, {PureComponent} from 'react'
 import _ from 'lodash'
 
-import getLastValues, {TimeSeriesResponse} from 'src/shared/parsing/lastValues'
+import getLastValues from 'src/shared/parsing/lastValues'
 import Gauge from 'src/shared/components/Gauge'
 
 import {DEFAULT_GAUGE_COLORS} from 'src/shared/constants/thresholds'
@@ -9,22 +9,15 @@ import {stringifyColorValues} from 'src/shared/constants/colorOperations'
 import {DASHBOARD_LAYOUT_ROW_HEIGHT} from 'src/shared/constants'
 import {ErrorHandling} from 'src/shared/decorators/errors'
 import {DecimalPlaces} from 'src/types/dashboards'
-
-interface Color {
-  type: string
-  hex: string
-  id: string
-  name: string
-  value: string
-}
+import {ColorString} from 'src/types/colors'
+import {TimeSeriesServerResponse} from 'src/types/series'
 
 interface Props {
-  data: TimeSeriesResponse[]
+  data: TimeSeriesServerResponse[]
   decimalPlaces: DecimalPlaces
-  isFetchingInitially: boolean
   cellID: string
   cellHeight?: number
-  colors?: Color[]
+  colors?: ColorString[]
   prefix: string
   suffix: string
   resizerTopHeight?: number
@@ -37,16 +30,7 @@ class GaugeChart extends PureComponent<Props> {
   }
 
   public render() {
-    const {isFetchingInitially, colors, prefix, suffix} = this.props
-
-    if (isFetchingInitially) {
-      return (
-        <div className="graph-empty">
-          <h3 className="graph-spinner" />
-        </div>
-      )
-    }
-
+    const {colors, prefix, suffix, decimalPlaces} = this.props
     return (
       <div className="single-stat">
         <Gauge
@@ -56,6 +40,7 @@ class GaugeChart extends PureComponent<Props> {
           prefix={prefix}
           suffix={suffix}
           gaugePosition={this.lastValueForGauge}
+          decimalPlaces={decimalPlaces}
         />
       </div>
     )
@@ -78,16 +63,12 @@ class GaugeChart extends PureComponent<Props> {
   }
 
   private get lastValueForGauge(): number {
-    const {data, decimalPlaces} = this.props
+    const {data} = this.props
     const {lastValues} = getLastValues(data)
-    let lastValue = _.get(lastValues, 0, 0)
+    const lastValue = _.get(lastValues, 0, 0)
 
     if (!lastValue) {
       return 0
-    }
-
-    if (decimalPlaces.isEnforced) {
-      lastValue = +lastValue.toFixed(decimalPlaces.digits)
     }
 
     return lastValue