+++ ui/src/tempVars/utils/replace.ts
@@ -1,3 +1,5 @@
+import {topologicalSort, graphFromTemplates} from 'src/tempVars/utils/graph'
+
 import {
   Template,
   TemplateType,
@@ -9,6 +11,12 @@ import {
   DEFAULT_PIXELS,
   DEFAULT_DURATION_MS,
 } from 'src/shared/constants'
+
+function sortTemplates(templates: Template[]): Template[] {
+  const graph = graphFromTemplates(templates)
+
+  return topologicalSort(graph).map(t => t.initialTemplate)
+}
 
 export const replaceInterval = (
   query: string,
@@ -33,27 +41,10 @@ export const replaceInterval = (
   return replaceAll(query, TEMP_VAR_INTERVAL, `${msPerPixel}ms`)
 }
 
-const TEMPLATES_SORTING_ORDER = {
-  [TemplateType.CSV]: 0,
-  [TemplateType.Map]: 0,
-  [TemplateType.AutoGroupBy]: 1,
-  [TemplateType.Constant]: 1,
-  [TemplateType.FieldKeys]: 1,
-  [TemplateType.Measurements]: 1,
-  [TemplateType.TagKeys]: 1,
-  [TemplateType.TagValues]: 1,
-  [TemplateType.Databases]: 1,
-  [TemplateType.MetaQuery]: 1,
-}
+const templateReplace = (query: string, templates: Template[]) => {
+  const sortedTemplates = sortTemplates(templates)
 
-const sortTemplates = (a: Template, b: Template): number => {
-  return TEMPLATES_SORTING_ORDER[a.type] - TEMPLATES_SORTING_ORDER[b.type]
-}
-
-const templateReplace = (query: string, tempVars: Template[]) => {
-  const sortedTempVars = [...tempVars].sort(sortTemplates)
-
-  return sortedTempVars.reduce(
+  return sortedTemplates.reduce(
     (acc, template) => renderTemplate(acc, template),
     query
   )
@@ -143,6 +134,23 @@ const replaceAllRegex = (
 
 const replaceAll = (query: string, search: string, replacement: string) => {
   return query.split(search).join(replacement)
+}
+
+export const templateInternalReplace = (template: Template): string => {
+  const {influxql, db, measurement, tagKey} = template.query
+
+  if (template.type === TemplateType.MetaQuery) {
+    // A custom meta query template may reference other templates whose names
+    // conflict with the `database`, `measurement` and `tagKey` fields stored
+    // within a template's `query` object. Since these fields are always empty
+    // for a custom meta query template, we do not attempt to replace them
+    return influxql
+  }
+
+  return influxql
+    .replace(':database:', `"${db}"`)
+    .replace(':measurement:', `"${measurement}"`)
+    .replace(':tagKey:', `"${tagKey}"`)
 }
 
 export default templateReplace