+++ ui/test/resources.ts
@@ -633,15 +633,6 @@ export const template: Template = {
   ],
 }
 
-export const emptyTemplate: Template = {
-  id: '1',
-  type: TemplateType.CSV,
-  label: '',
-  tempVar: '',
-  query: {},
-  values: [],
-}
-
 export const dashboard: Dashboard = {
   id: 1,
   cells: [],
@@ -698,4 +689,5 @@ export const cell: Cell = {
       '/chronograf/v1/dashboards/10/cells/8b3b7897-49b1-422c-9443-e9b778bcbf12',
   },
   legend: {},
+  inView: true,
 }